+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/statements/domain/model/ClassService.kt
@@ -1,5 +1,0 @@
-package eu.tib.orkg.prototype.statements.domain.model
-
-import eu.tib.orkg.prototype.statements.api.ClassUseCases
-
-interface ClassService : ClassUseCases