+++ rest-api-server/src/test/kotlin/eu/tib/orkg/prototype/statements/application/ClassControllerTest.kt
@@ -2,16 +2,23 @@ package eu.tib.orkg.prototype.statements.application
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.ninjasquad.springmockk.MockkBean
+import dev.forkhandles.result4k.Failure
+import dev.forkhandles.result4k.Success
 import eu.tib.orkg.prototype.AuthorizationServerUnitTestWorkaround
 import eu.tib.orkg.prototype.auth.service.UserRepository
+import eu.tib.orkg.prototype.createClass
+import eu.tib.orkg.prototype.statements.api.AlreadyInUse
 import eu.tib.orkg.prototype.statements.api.ClassUseCases
+import eu.tib.orkg.prototype.statements.api.InvalidURI
+import eu.tib.orkg.prototype.statements.api.UpdateNotAllowed
 import eu.tib.orkg.prototype.statements.api.ResourceUseCases
 import eu.tib.orkg.prototype.statements.domain.model.Class
 import eu.tib.orkg.prototype.statements.domain.model.ClassId
 import io.mockk.every
+import io.mockk.verify
 import java.net.URI
 import java.time.OffsetDateTime
-import java.util.Optional
+import java.util.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
@@ -20,19 +27,27 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.http.MediaType
 import org.springframework.test.web.servlet.MockMvc
+import org.springframework.test.web.servlet.ResultActions
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.content
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 import org.springframework.test.web.servlet.setup.MockMvcBuilders
 import org.springframework.web.context.WebApplicationContext
+import eu.tib.orkg.prototype.statements.api.ClassNotFound as ClassNotFoundProblem
+import eu.tib.orkg.prototype.statements.api.InvalidLabel as InvalidLabelProblem
+
+internal const val INVALID_LABEL = "invalid\nlabel"
+internal const val INVALID_URI = "invalid\nuri"
 
 @WebMvcTest(controllers = [ClassController::class])
 @AuthorizationServerUnitTestWorkaround
 @DisplayName("Given a Class controller")
-class ClassControllerTest {
+internal class ClassControllerTest {
 
     private lateinit var mockMvc: MockMvc
 
@@ -107,6 +122,202 @@ class ClassControllerTest {
                 .andExpect(requestContentIsEmpty())
         }
     }
+
+    @Test
+    fun `Given the class is replaced, when service succeeds, then status is 200 OK and class is returned`() {
+        val id = ClassId("EXISTS")
+        val replacingClass = createClass().copy(label = "new label")
+        val body = objectMapper.writeValueAsString(replacingClass)
+        every { classService.replace(id, with = any()) } returns Success(Unit)
+
+        mockMvc.performPut("/api/classes/$id", body).andExpect(status().isOk)
+            .andExpect(jsonPath("\$.id").value(id.toString()))
+        verify(exactly = 1) { classService.replace(any(), any()) }
+    }
+
+    @Test
+    fun `Given the class is replaced, when service reports class cannot be found, then status is 404 NOT FOUND`() {
+        val id = ClassId("NON-EXISTENT")
+        val replacingClass = createClass().copy(label = "new label")
+        val body = objectMapper.writeValueAsString(replacingClass)
+        every { classService.replace(id, with = any()) } returns Failure(ClassNotFoundProblem)
+
+        mockMvc.performPut("/api/classes/$id", body).andExpect(status().isNotFound)
+            .andExpect(requestContentIsEmpty())
+    }
+
+    @Test
+    fun `Given the class is replaced, when service reports label is invalid, then status is 400 BAD REQUEST and returns error information`() {
+        val id = ClassId("EXISTS")
+        val replacingClass = createClass().copy(label = INVALID_LABEL)
+        val body = objectMapper.writeValueAsString(replacingClass)
+        every { classService.replace(id, with = any()) } returns Failure(InvalidLabelProblem)
+
+        mockMvc.performPut("/api/classes/$id", body)
+            .andExpect(status().isBadRequest)
+            .andExpect(jsonPath("\$.status").value(400))
+            .andExpect(jsonPath("\$.errors.length()").value(1))
+            .andExpect(jsonPath("\$.errors[0].field").value("label"))
+            .andExpect(jsonPath("\$.errors[0].message").value("A label must not be blank or contain newlines."))
+    }
+
+    @Test
+    fun `Given the class is replaced, when service reports changing URI is not allowed, then status is 403 FORBIDDEN and returns error information`() {
+        val id = ClassId("EXISTS")
+        val replacingClass = createClass().copy(label = "new label", uri = URI.create("https://example.com/NEW#uri"))
+        val body = objectMapper.writeValueAsString(replacingClass)
+        every { classService.replace(id, with = any()) } returns Failure(UpdateNotAllowed)
+
+        mockMvc.performPut("/api/classes/$id", body)
+            .andExpect(status().isForbidden)
+            .andExpect(jsonPath("\$.status").value(403))
+            .andExpect(jsonPath("\$.errors.length()").value(1))
+            .andExpect(jsonPath("\$.errors[0].field").value("uri"))
+            .andExpect(jsonPath("\$.errors[0].message").value("The class EXISTS already has a URI. It is not allowed to change URIs."))
+    }
+
+    @Test
+    fun `Given the class is replaced, when service reports URI is in use, then status is 403 FORBIDDEN and returns error information`() {
+        val id = ClassId("EXISTS")
+        val replacingClass = createClass().copy(label = "new label", uri = URI.create("https://example.com/NEW#uri"))
+        val body = objectMapper.writeValueAsString(replacingClass)
+        every { classService.replace(id, with = any()) } returns Failure(AlreadyInUse)
+
+        mockMvc.performPut("/api/classes/$id", body)
+            .andExpect(status().isForbidden)
+            .andExpect(jsonPath("\$.status").value(403))
+            .andExpect(jsonPath("\$.errors.length()").value(1))
+            .andExpect(jsonPath("\$.errors[0].field").value("uri"))
+            .andExpect(jsonPath("\$.errors[0].message").value("The URI <https://example.com/NEW#uri> is already in use by another class."))
+    }
+
+    @Test
+    fun `Given the class label is patched, when service reports class cannot be found, then status is 404 NOT FOUND`() {
+        val id = ClassId("NON-EXISTENT")
+        val body = mapOf("label" to "some label")
+        every {
+            classService.updateLabel(id, "some label")
+        } returns Failure(ClassNotFoundProblem)
+
+        mockMvc.performPatch("/api/classes/$id", body).andExpect(status().isNotFound)
+            .andExpect(requestContentIsEmpty())
+    }
+
+    @Test
+    fun `Given the class label is patched, when service succeeds, then status is 200 OK`() {
+        val id = ClassId("EXISTS")
+        val body = mapOf("label" to "some label")
+        every { classService.updateLabel(id, "some label") } returns Success(Unit)
+
+        mockMvc.performPatch("/api/classes/$id", body).andExpect(status().isOk)
+    }
+
+    @Test
+    fun `Given the class label is patched and no label is provided, then status is 200 OK and no action was taken`() {
+        val dummyId = "EXISTS"
+        val body = mapOf("label" to null)
+
+        mockMvc.performPatch("/api/classes/$dummyId", body).andExpect(status().isOk)
+        verify(exactly = 0) { classService.updateLabel(any(), any()) }
+    }
+
+    @Test
+    fun `Given the class label is patched, when service reports the label is invalid, then status is 400 BAD REQUEST and returns error information`() {
+        val id = ClassId("EXISTS")
+        val body = mapOf("label" to INVALID_LABEL)
+        every { classService.updateLabel(id, INVALID_LABEL) } returns Failure(InvalidLabelProblem)
+
+        mockMvc.performPatch("/api/classes/$id", body)
+            .andExpect(status().isBadRequest)
+            .andExpect(jsonPath("\$.status").value(400))
+            .andExpect(jsonPath("\$.errors.length()").value(1))
+            .andExpect(jsonPath("\$.errors[0].field").value("label"))
+            .andExpect(jsonPath("\$.errors[0].message").value("A label must not be blank or contain newlines."))
+    }
+
+    @Test
+    fun `Given the class URI is patched, when service reports class cannot be found, then status is 404 NOT FOUND`() {
+        val id = ClassId("NON-EXISTENT")
+        val body = mapOf("uri" to "https://example.org")
+        every {
+            classService.updateURI(
+                id, "https://example.org"
+            )
+        } returns Failure(ClassNotFoundProblem)
+
+        mockMvc.performPatch("/api/classes/$id", body).andExpect(status().isNotFound)
+            .andExpect(requestContentIsEmpty())
+    }
+
+    @Test
+    fun `Given the class URI is patched and a valid URI is provided, when service succeeds, then status is 200 OK`() {
+        val id = ClassId("EXISTS")
+        val body = mapOf("uri" to "https://example.org/some/new#URI")
+        every { classService.updateURI(id, "https://example.org/some/new#URI") } returns Success(Unit)
+
+        mockMvc.performPatch("/api/classes/$id", body).andExpect(status().isOk)
+        verify(exactly = 1) { classService.updateURI(any(), any()) }
+    }
+
+    @Test
+    fun `Given the class URI is patched, when service reports updating is not allowed, then status is 403 FORBIDDEN and returns error information`() {
+        val id = ClassId("EXISTS")
+        val body = mapOf("uri" to INVALID_URI)
+        every { classService.updateURI(id, INVALID_URI) } returns Failure(UpdateNotAllowed)
+
+        mockMvc.performPatch("/api/classes/$id", body)
+            .andExpect(status().isForbidden)
+            .andExpect(jsonPath("\$.status").value(403))
+            .andExpect(jsonPath("\$.errors.length()").value(1))
+            .andExpect(jsonPath("\$.errors[0].field").value("uri"))
+            .andExpect(jsonPath("\$.errors[0].message").value("The class EXISTS already has a URI. It is not allowed to change URIs."))
+    }
+
+    @Test
+    fun `Given the class URI is patched, when service reports URI is in use, then status is 403 FORBIDDEN and returns error information`() {
+        val id = ClassId("EXISTS")
+        val body = mapOf("uri" to "https://example.org/some/new#URI")
+        every { classService.updateURI(id, "https://example.org/some/new#URI") } returns Failure(AlreadyInUse)
+
+        mockMvc.performPatch("/api/classes/$id", body)
+            .andExpect(status().isForbidden)
+            .andExpect(jsonPath("\$.status").value(403))
+            .andExpect(jsonPath("\$.errors.length()").value(1))
+            .andExpect(jsonPath("\$.errors[0].field").value("uri"))
+            .andExpect(jsonPath("\$.errors[0].message").value("The URI <https://example.org/some/new#URI> is already in use by another class."))
+    }
+
+    @Test
+    fun `Given the class URI is patched, when service reports URI is invalid, then status is 400 BAD REQUEST and returns error information`() {
+        val id = ClassId("EXISTS")
+        val body = mapOf("uri" to INVALID_URI)
+        every { classService.updateURI(id, INVALID_URI) } returns Failure(InvalidURI)
+
+        mockMvc.performPatch("/api/classes/$id", body)
+            .andExpect(status().isBadRequest)
+            .andExpect(jsonPath("\$.status").value(400))
+            .andExpect(jsonPath("\$.errors.length()").value(1))
+            .andExpect(jsonPath("\$.errors[0].field").value("uri"))
+            .andExpect(jsonPath("\$.errors[0].message").value("The provided URI is not a valid URI."))
+    }
+
+    @Test
+    fun `Given the class URI is patched and no URI is provided, then status is 200 OK and no action was taken`() {
+        val dummyId = "EXISTS"
+        val body = mapOf("uri" to null)
+
+        mockMvc.performPatch("/api/classes/$dummyId", body).andExpect(status().isOk)
+        verify(exactly = 0) { classService.updateURI(any(), any()) }
+    }
+
+    private fun MockMvc.performPatch(urlTemplate: String, body: Map<String, Any?>): ResultActions = perform(
+        patch(urlTemplate).contentType(MediaType.APPLICATION_JSON).characterEncoding("UTF-8")
+            .content(objectMapper.writeValueAsString(body))
+    )
+
+    private fun MockMvc.performPut(urlTemplate: String, body: String): ResultActions = perform(
+        put(urlTemplate).contentType(MediaType.APPLICATION_JSON).characterEncoding("UTF-8").content(body)
+    )
 
     private fun requestContentIsEmpty() = content().string("")